ai/requirements.txt
# =============================================================================
# OPTIMIZED REQUIREMENTS.TXT - Grouped by installation speed
# =============================================================================

# -----------------------------------------------------------------------------
# GROUP 1: Fast-installing packages (install first for quick feedback)
# -----------------------------------------------------------------------------
# Core utilities
numpy>=1.24.0
pandas>=2.0.0

# Web frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration & Templates
jinja2>=3.1.0
PyYAML>=6.0

# RPC
grpcio>=1.59.0
grpcio-tools>=1.59.0

# Caching
redis>=5.0.0

# Logging & Monitoring
loguru>=0.7.0
tqdm>=4.66.0

# Hot reload
watchdog>=6.0.0
# -----------------------------------------------------------------------------
# GROUP 2: Medium installation time
# -----------------------------------------------------------------------------
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# GROUP 3: Heavy packages (PyTorch ecosystem) - Install from pre-built wheels
# -----------------------------------------------------------------------------
# NOTE: torch, torchvision, torchaudio are installed separately via:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# This is MUCH faster than installing from PyPI

# PyTorch utilities (install after torch)
pytorch-lightning>=2.1.0

# Transformers ecosystem (large but cached well)
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0

# -----------------------------------------------------------------------------
# GROUP 4: Optional - Comment out if not needed
# -----------------------------------------------------------------------------
# Experiment tracking (only if using Weights & Biases)
wandb>=0.16.0

# Hyperparameter optimization (only if needed)
optuna>=3.4.0

# -----------------------------------------------------------------------------
# Production-only dependencies (uncomment if needed)
# -----------------------------------------------------------------------------
# prometheus-client>=0.19.0  # Metrics
# sentry-sdk>=1.38.0         # Error tracking
# python-jose[cryptography]  # JWT tokens